+++ stable_baselines/acer/acer_simple.py
@@ -566,9 +566,9 @@ class ACER(ActorCriticRLModel):
             "policy_kwargs": self.policy_kwargs
         }
 
-        params = self.sess.run(self.params)
+        params_to_save = self.get_parameters()
 
-        self._save_to_file(save_path, data=data, params=params)
+        self._save_to_file(save_path, data=data, params=params_to_save)
 
 
 class _Runner(AbstractEnvRunner):