+++ stable_baselines/ppo2/ppo2.py
@@ -380,9 +380,9 @@ class PPO2(ActorCriticRLModel):
             "policy_kwargs": self.policy_kwargs
         }
 
-        params = self.sess.run(self.params)
+        params_to_save = self.get_parameters()
 
-        self._save_to_file(save_path, data=data, params=params)
+        self._save_to_file(save_path, data=data, params=params_to_save)
 
 
 class Runner(AbstractEnvRunner):